Makefile: Add Makefile for easy installation and uninstallation
The mkpki source has neither an installation script nor instructions
how to install it, which means that users will either have to install
the scripts manually or use them without installing them first.

This commit adds a Makefile, allowing the scripts to be installed and
uninstalled by running `make install` and `make uninstall`.