Add README with usage examples
Since it is not obvious how mkpki has to be used, and since there is
no readme in the repository, new users won't be able to easily use
mkpki.

This commit adds a readme explaining the purpose of mkpki and how to
use the commands that are part of it.